refactor: Main 액티비티를 Home 액티비티로 변경
- Main이 아닌 Home으로 변경함
