+++ app/src/main/AndroidManifest.xml
@@ -29,7 +29,7 @@
         </activity>
 
         <activity
-            android:name=".presentation.main.MainActivity"
+            android:name=".presentation.home.HomeActivity"
             android:exported="false" />
     </application>
